backend/shadowsec-dashboard/backend/requirements.txt
fastapi
uvicorn
python-multipart
requests
pyjwt
